Update launcher example
