+++ python/launcher/client.py
@@ -1,35 +1,23 @@
-#!/usr/bin/env python
-#
-# example code to test launcher
-#
-# usage:
-#   LS_ACCESS_TOKEN=${SECRET_TOKEN} \
-#   LS_SERVICE_NAME=demo-python \
-#   LS_SERVICE_VERSION=0.0.8 \
-#   opentelemetry-instrument python client.py
+from sys import argv
 
-import os
-import time
+from requests import get
 
 from opentelemetry import trace
+from opentelemetry.propagate import inject
 
-import requests
+tracer = trace.get_tracer_provider().get_tracer(__name__)
 
-tracer = trace.get_tracer(__name__)
+assert len(argv) == 2
 
+with tracer.start_as_current_span("client"):
 
-def send_requests(url):
-    with tracer.start_as_current_span("client operation"):
-        try:
-            res = requests.get(url)
-            print(f"Request to {url}, got {len(res.content)} bytes")
-        except Exception as e:
-            print(f"Request to {url} failed {e}")
-            pass
+    with tracer.start_as_current_span("client-server"):
+        headers = {}
+        inject(headers)
+        requested = get(
+            "http://localhost:8082/rolldice",
+            params={"param": argv[1]},
+            headers=headers,
+        )
 
-
-if __name__ == "__main__":
-    target = os.getenv("DESTINATION_URL", "http://localhost:8081/ping")
-    while True:
-        send_requests(target)
-        time.sleep(5)
+        assert requested.status_code == 200
