+++ python/launcher/requirements.txt
@@ -1,10 +1,5 @@
-protobuf==3.20.1
-flask
-pymongo
-redis
+# OTel-specific
+opentelemetry-launcher
 requests
-sqlalchemy
-opentelemetry-instrumentation==0.30b1
-opentelemetry-launcher==1.8.0
-opentelemetry-instrumentation-requests
-opentelemetry-instrumentation-flask
+flask
+protobuf==3.20.1
