--- python/opentelemetry/otlp/app2.py
@@ -1,26 +1,0 @@
-# These are the necessary import declarations
-from opentelemetry import trace
-
-from random import randint
-from flask import Flask, request
-
-# Acquire a tracer
-tracer = trace.get_tracer(__name__)
-
-app = Flask(__name__)
-
-@app.route("/rolldice")
-def roll_dice():
-    print("Yo yo yo")
-    return str(do_roll())
-
-def do_roll():
-    # This creates a new span that's the child of the current one
-    with tracer.start_as_current_span("do_roll") as rollspan:  
-        res = randint(1, 6)
-        rollspan.set_attribute("roll.value", res)
-        return res
-    
-
-if __name__ == "__main__":
-    app.run(host="0.0.0.0", port=6000)
