--- python/opentelemetry/otlp/requirements.txt
@@ -1,10 +1,0 @@
-flask
-uuid
-opentelemetry-api
-opentelemetry-sdk
-opentelemetry-exporter-otlp
-opentelemetry-exporter-otlp-proto-http
-opentelemetry-exporter-otlp-proto-grpc
-
-opentelemetry-instrumentation-flask
-opentelemetry-instrumentation-requests
